Added FigmaAssetProcessor
