+++ lib/input/figma/entities/layers/boolean_operation.dart
@@ -4,20 +4,18 @@ import 'package:parabeac_core/design_logic/image.dart';
 import 'package:parabeac_core/input/figma/entities/abstract_figma_node_factory.dart';
 import 'package:parabeac_core/input/figma/entities/layers/vector.dart';
 import 'package:parabeac_core/input/figma/entities/style/figma_style.dart';
+import 'package:parabeac_core/input/figma/helper/figma_asset_processor.dart';
 import 'package:parabeac_core/input/sketch/entities/objects/frame.dart';
 import 'package:parabeac_core/interpret_and_optimize/entities/inherited_bitmap.dart';
 import 'package:parabeac_core/interpret_and_optimize/entities/subclasses/pb_intermediate_node.dart';
 import 'package:parabeac_core/interpret_and_optimize/helpers/pb_context.dart';
 import 'package:quick_log/quick_log.dart';
 import 'figma_node.dart';
-import 'package:parabeac_core/input/figma/helper/image_helper.dart'
-    as image_helper;
 
 part 'boolean_operation.g.dart';
 
 @JsonSerializable(nullable: true)
 class BooleanOperation extends FigmaVector
-    with image_helper.PBImageHelperMixin
     implements FigmaNodeFactory, GroupNode, Image {
   @JsonKey(ignore: true)
   Logger log;
@@ -55,7 +53,7 @@ class BooleanOperation extends FigmaVector
 
   @override
   Future<PBIntermediateNode> interpretNode(PBContext currentContext) async {
-    imageReference = addToImageQueue(UUID);
+    imageReference = FigmaAssetProcessor().processImage(UUID);
 
     return Future.value(
         InheritedBitmap(this, name, currentContext: currentContext));
