+++ lib/input/figma/entities/layers/ellipse.dart
@@ -1,21 +1,19 @@
 import 'package:parabeac_core/design_logic/image.dart';
 import 'package:parabeac_core/input/figma/entities/abstract_figma_node_factory.dart';
 import 'package:parabeac_core/input/figma/entities/layers/vector.dart';
+import 'package:parabeac_core/input/figma/helper/figma_asset_processor.dart';
 import 'package:parabeac_core/input/sketch/entities/objects/frame.dart';
 import 'package:parabeac_core/interpret_and_optimize/entities/inherited_bitmap.dart';
 import 'package:parabeac_core/interpret_and_optimize/entities/subclasses/pb_intermediate_node.dart';
 import 'package:parabeac_core/interpret_and_optimize/helpers/pb_context.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:quick_log/quick_log.dart';
-import 'package:parabeac_core/input/figma/helper/image_helper.dart'
-    as image_helper;
 import 'figma_node.dart';
 
 part 'ellipse.g.dart';
 
 @JsonSerializable(nullable: true)
 class FigmaEllipse extends FigmaVector
-    with image_helper.PBImageHelperMixin
     implements AbstractFigmaNodeFactory, Image {
   @override
   String imageReference;
@@ -72,7 +70,7 @@ class FigmaEllipse extends FigmaVector
 
   @override
   Future<PBIntermediateNode> interpretNode(PBContext currentContext) async {
-    imageReference = addToImageQueue(UUID);
+    imageReference = FigmaAssetProcessor().processImage(UUID);
     return Future.value(
         InheritedBitmap(this, name, currentContext: currentContext));
   }
