+++ lib/input/figma/entities/layers/frame.dart
@@ -14,7 +14,7 @@ import 'package:parabeac_core/interpret_and_optimize/helpers/pb_context.dart';
 import 'package:parabeac_core/interpret_and_optimize/entities/subclasses/pb_intermediate_node.dart';
 import 'package:parabeac_core/input/sketch/entities/objects/frame.dart';
 import 'package:parabeac_core/interpret_and_optimize/value_objects/point.dart';
-import 'package:parabeac_core/input/figma/helper/image_helper.dart'
+import 'package:parabeac_core/input/figma/helper/figma_asset_processor.dart'
     as image_helper;
 
 part 'frame.g.dart';
