+++ lib/input/figma/entities/layers/group.dart
@@ -7,8 +7,7 @@ import 'package:parabeac_core/input/figma/entities/layers/frame.dart';
 import 'package:parabeac_core/input/figma/entities/layers/text.dart';
 import 'package:parabeac_core/input/figma/entities/layers/vector.dart';
 import 'package:parabeac_core/input/figma/entities/style/figma_color.dart';
-import 'package:parabeac_core/input/figma/helper/image_helper.dart'
-    as image_helper;
+import 'package:parabeac_core/input/figma/helper/figma_asset_processor.dart';
 import 'package:parabeac_core/input/sketch/entities/layers/flow.dart';
 import 'package:parabeac_core/input/sketch/entities/objects/frame.dart';
 import 'package:parabeac_core/interpret_and_optimize/entities/inherited_bitmap.dart';
@@ -25,9 +24,7 @@ part 'group.g.dart';
 /// Class that represents a Figma Group.
 /// The reason this class implements Image is because Groups can hold multiple vectors
 /// which we need to convert into images.
-class Group extends FigmaFrame
-    with image_helper.PBImageHelperMixin
-    implements AbstractFigmaNodeFactory, Image {
+class Group extends FigmaFrame implements AbstractFigmaNodeFactory, Image {
   @JsonKey(ignore: true)
   Logger log;
   @override
@@ -95,7 +92,7 @@ class Group extends FigmaFrame
   @override
   Future<PBIntermediateNode> interpretNode(PBContext currentContext) async {
     if (areAllVectors()) {
-      imageReference = addToImageQueue(UUID);
+      imageReference = FigmaAssetProcessor().processImage(UUID);
 
       var tempPrototypeID = childrenHavePrototypeNode();
       if (tempPrototypeID != null) {
